[VT]: Fixed VT server change string value bug, add a way to defer sending the load version response
Fixed an issue where we were supposed to be padding VT client's change string commands
with spaces if they sent us a string that was too short.
Added a way to defer sending the load version response until the parsing
thread is joined. This prevents a race condition against the client commanding
us to do something, and the thread completing its parsing.